backend/requirements.txt
fastapi
uvicorn
python-jose
python-multipart
firebase-admin
pydantic
